% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNoDataModels.R
\name{addNoDataModels}
\alias{addNoDataModels}
\title{Add Ensemble Weights for Sets of the Data without Valid Observations}
\usage{
addNoDataModels(data, weights, imputationParameters)
}
\arguments{
\item{data}{The data.table object containing the data which will be imputed.}

\item{weights}{A data.table containing the currently estimated ensemble
weights and errors.  These are used to compute the global errors and add new
weights for models with missing data.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\value{
A data.table of the same format as weights.  This function is
really just a helper function for getEnsembleWeight, and so the output of
the function is specifically designed to be rbinded on to the weights
object in that function.
}
\description{
In some cases, countries (or whatever the byKey is) will not have any data
available.  In these cases, we can impute values for these observations via
an ensemble, but we'll need a method for weighting the models in the
ensemble.
}
\details{
This function computes the weights for these models by looking at the
cross-validation error of the global models and uses those errors to
determine the final weights.
}

