% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allDefaultModels.R
\name{allDefaultModels}
\alias{allDefaultModels}
\title{Returns all default models}
\usage{
allDefaultModels()
}
\value{
A list of all the default model functions.
}
\description{
This is a convenience function that returns a list of all the default
models.
}

