% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class-ensembleModel.R
\name{checkEnsembleModel}
\alias{checkEnsembleModel}
\title{Check Ensemble Model}
\usage{
checkEnsembleModel(object)
}
\arguments{
\item{object}{An S4 object.}
}
\value{
TRUE if there are no errors, otherwise the error messages.
}
\description{
Check Ensemble Model
}

