% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEnsemble.R
\name{computeEnsemble}
\alias{computeEnsemble}
\title{Function to combine the ensembles}
\usage{
computeEnsemble(fits, weights, errors)
}
\arguments{
\item{fits}{A list of fitted values from models.}

\item{weights}{A weight matrix giving the weights of each model for each
observation.}

\item{errors}{An error matrix giving the error of each model for each
observation.  This is used to create an estimate of variability of the
ensemble for each imputed value.}
}
\value{
A vector of length nrow(weights), where each value represents the
ensemble estimate for that observation.
}
\description{
Function to combine the ensembles
}

