% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeEnsembleWeight.R
\name{computeEnsembleWeight}
\alias{computeEnsembleWeight}
\title{Function to compute the weights of the ensemble models}
\usage{
computeEnsembleWeight(data, cvGroup, fits, method = "inverse",
  imputationParameters)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{cvGroup}{A vector of the same length as nrow(data).  Entries of the
vector should be integers from 1 to the number of cross-validation groups
(typically 10).  This should be randomly assigned, and is usually created
by ensembleImpute.}

\item{fits}{The fitted values from the models.}

\item{method}{Must be either "inverse" or "stacking".  If "inverse", the
final ensemble is a weighted average of all the individual models, where
the weight of each model is proportional to 1/error from that model.  If
"stacking", then the weight is assigned via a linear regression (where the
independent variable in the regression is the variable being imputed, and
each individual model is a dependent variable).  The linear regression is
restricted, however: no weights may be negative and the weights must sum to
one.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\description{
Function to compute the weights of the ensemble models
}

