% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/computeErrorLOOCV.R
\name{computeErrorLOOCV}
\alias{computeErrorLOOCV}
\title{LOOCV Error Rate}
\usage{
computeErrorLOOCV(data, model, cvGroup, imputationParameters)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{model}{The model fit to x, should be of class ensembleModel.}

\item{cvGroup}{A vector of the same length as nrow(data).  Entries of the
vector should be integers from 1 to the number of cross-validation groups
(typically 10).  This should be randomly assigned, and is usually created
by ensembleImpute.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\description{
This function computes the error of different model fits via leave-one-out
cross-validation.  However, typically this function will be called via
computeErrorRate and not directly.
}

