% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultExp.R
\name{defaultExp}
\alias{defaultExp}
\title{The default exponential model for the ensemble model.}
\usage{
defaultExp(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default exponential model for the ensemble model.
}

