% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultGlobalMean.R
\name{defaultGlobalMean}
\alias{defaultGlobalMean}
\title{Global Mean Model for Imputation}
\usage{
defaultGlobalMean(data, imputationParameters)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\value{
Returns a vector of the estimated/imputed values.  If a value
existed in the original data, then an NA is returned in that location.
Otherwise, the mean of all available observations is returned (computed
across the entire dataset).
}
\description{
This function imputes missing values through a global mean.
}

