% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultLoess.R
\name{defaultLoess}
\alias{defaultLoess}
\title{The default LOESS model for the ensemble model.}
\usage{
defaultLoess(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default LOESS model for the ensemble model.
}

