% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultMars.R
\name{defaultMars}
\alias{defaultMars}
\title{The default MARS model for the ensemble model.}
\usage{
defaultMars(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default MARS model for the ensemble model.
}

