% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultMedian.R
\name{defaultMedian}
\alias{defaultMedian}
\title{The default median model for the ensemble model.}
\usage{
defaultMedian(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}
}
\description{
The default median model for the ensemble model.
}

