% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaultMovingAverage.R
\name{defaultMovingAverage}
\alias{defaultMovingAverage}
\title{The default moving average model.}
\usage{
defaultMovingAverage(x)
}
\arguments{
\item{x}{A numeric vector to be imputed.}

\item{lookback}{The number of previous observations to use.}
}
\description{
This module is very simple: it uses a mean of the last k observed (non-NA)
values to predict the next value.  k is specified by "lookback".  If there
aren't k previous values, the number available are used.  If there aren't
any previous values, the global mean is imputed (imputing NA could cause
issues with the ensemble weighting process).
}
\examples{
defaultMovingAverage(x = c(10, 10, 10, NA, 1, NA))
defaultMovingAverage(x = c(NA, 1:5, NA))
defaultMovingAverage(x = c(NA, 1, NA, 2, NA, 3, NA, 4, NA, 5, NA))
}

