% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensembleImpute.R
\name{ensembleImpute}
\alias{ensembleImpute}
\title{Function to perform ensemble imputation}
\usage{
ensembleImpute(data, imputationParameters)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\description{
This is an implementation of the ensemble imputation methodology
developed for the FAO production domain.
}

