% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Class-ensembleModel.R
\docType{class}
\name{ensembleModel-class}
\alias{ensembleModel}
\alias{ensembleModel-class}
\title{Ensemble Model Class}
\description{
\describe{
    \item{\code{model}:}{The function defining how the model is fit to
    the data.  The function should take two arguments: data and
    imputationParameters (see extendSimpleModel)}.
    \item{\code{extrapolationRange}:}{How many time steps outside of
    the data is this model valid for?  Should be a positive integer
    (or Inf).}
    \item{\code{level}:}{The level at which this model is applied.
    Currently, must be one of "countryCommodity" or "commodity".  This
    defines if this model operates on each country-commodity pair
    individually ("countryCommodity" level) or all countries for a fixed
    commodity ("commodity" level) at once.}
}
}

