% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faoswsImputation-package.R
\docType{package}
\name{faoswsImputation-package}
\alias{faoswsImputation}
\alias{faoswsImputation-package}
\title{Package to perform ensemble imputation for various FAO domains.}
\description{
Repository for storing code and associated files for the imputation package.
This package implements flexible data imputation with the use of ensembles
of models.
}
\author{
Joshua M. Browning \email{joshua.browning@fao.org}, Michael. C. J.
  Kao \email{michael.kao@fao.org}
}
\keyword{package}

