% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovariates.R
\name{getCovariates}
\alias{getCovariates}
\title{Get Covariates}
\usage{
getCovariates(data, imputationParameters)
}
\arguments{
\item{data}{The data.table object where imputation is to be performed.
This object is used to determine which countries and times should be
queried from the World Bank dataset.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\value{
A data.table containing columns geographicAreaM49 and
timePointYears corresponding to the values in data as well as columns
containing World Bank data, such as GPD, paved road %, temperature, and
precipitation.
}
\description{
This function pulls data from the World Bank datasets, such as GDP
indicators and other such data.  These kinds of variables can be useful
for estimating production, loss, etc. and thus form a good set of covariate
variables for imputation models.
}

