% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeSingleObservation.R
\name{imputeSingleObservation}
\alias{imputeSingleObservation}
\title{Make Cross-Validation Groups}
\usage{
imputeSingleObservation(data, imputationParameters)
}
\arguments{
\item{data}{A data.table object containing the data.}

\item{imputationParameters}{A list of the parameters for the imputation
algorithms.  See defaultImputationParameters() for a starting point.}
}
\value{
A new data table with time series containing sinlge
    observation imputed.
}
\description{
Creates a vector of cross-validation groups to be used for leave-one-out
cross-validation in later models.
}

