% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputeVariable.R
\name{imputeVariable}
\alias{imputeVariable}
\title{Function to impute production or yield}
\usage{
imputeVariable(data, imputationParameters)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{imputationParameters}{A list of the parameters for the imputation
  algorithms.  See defaultImputationParameters() for a starting point.}
}
\value{
This function doesn't return any objects but modifies the underlying
  data.table that it was passed.
}
\description{
This is a wrapper of the ensemble imputation for the production domain.
}

