% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logisticGlm.R
\name{logisticGlm}
\alias{logisticGlm}
\title{Helper function for defaultLogistic}
\usage{
logisticGlm(x)
}
\arguments{
\item{x}{The dependent variable.}
}
\value{
A numeric vector of the same length as x but with the model
estimates at each point in time.  Note: if the logistic model has it's
midpoint outside the range of the data, this function will return a vector
of NA's (as the original defaultLogistic function had this behavior to
prevent poor fitting).
}
\description{
This function fits a logistic model to the data via logistic regression
(see ?glm and the binomial family).  To ensure all x values are between
0 and 1, the values are all divided by the maximum x value.  Thus, this
model assumes the asymptote is the largest observed x value.  Time is the
independent variable and is the vector 1:length(x).
}
\seealso{
Other logistic.functions: \code{\link{logisticNlsIntercept}};
  \code{\link{logisticNls}}
}

