% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okra.R
\docType{data}
\name{okra}
\alias{okra}
\title{Example data for the documentations.}
\format{A csv file with 912 rows and 14 variables}
\usage{
data(okra)
}
\description{
The data containing global okra production, area harvested and
yield from the year 1995 to 2013.
}
\keyword{datasets}

