% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDefaultRange.R
\name{getDefaultRange}
\alias{getDefaultRange}
\title{Default range for ensemble models}
\usage{
getDefaultRange(ensembleModel)
}
\arguments{
\item{ensembleModel}{A list of ensemble models to return ranges for.}
}
\value{
A numeric vector of the default ranges for each of the ensemble
models.
}
\description{
This function returns the default extrapolation range values for each model
in the input list of models.
}

